pyproject.toml
[project]
name = "py-playwright"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anyio==4.12.1",
    "certifi==2026.2.25",
    "greenlet==3.3.2",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.11",
    "lxml==6.0.2",
    "numpy==2.4.2",
    "openai>=2.26.0",
    "pandas==3.0.1",
    "playwright==1.58.0",
    "pyee==13.0.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv>=1.2.2",
    "six==1.17.0",
    "typing-extensions==4.15.0",
    "tzdata==2025.3",
]
